Renaming project files.
