--- Mock/Mock.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <EmitCompilerGeneratedFiles>true</EmitCompilerGeneratedFiles>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\MongoDbAccess.Generators\MongoDbAccess.Generators.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
-    <ProjectReference Include="..\MongoDbAccess\MongoDbAccess.csproj" />
-  </ItemGroup>
-	
-</Project>
