--- MongoDbAccess.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31112.23
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MongoDbAccess", "MongoDbAccess\MongoDbAccess.csproj", "{A14105D1-0E83-4F35-B6CD-A456280DFA28}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mock", "Mock\Mock.csproj", "{311CC301-5101-4F16-A0AA-5EA742BFF644}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MongoDbAccess.Generators", "MongoDbAccess.Generators\MongoDbAccess.Generators.csproj", "{8684E289-965E-41B1-9E3D-01B0C982B88C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A14105D1-0E83-4F35-B6CD-A456280DFA28}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A14105D1-0E83-4F35-B6CD-A456280DFA28}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A14105D1-0E83-4F35-B6CD-A456280DFA28}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A14105D1-0E83-4F35-B6CD-A456280DFA28}.Release|Any CPU.Build.0 = Release|Any CPU
-		{311CC301-5101-4F16-A0AA-5EA742BFF644}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{311CC301-5101-4F16-A0AA-5EA742BFF644}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{311CC301-5101-4F16-A0AA-5EA742BFF644}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{311CC301-5101-4F16-A0AA-5EA742BFF644}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8684E289-965E-41B1-9E3D-01B0C982B88C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8684E289-965E-41B1-9E3D-01B0C982B88C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8684E289-965E-41B1-9E3D-01B0C982B88C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8684E289-965E-41B1-9E3D-01B0C982B88C}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E6F448C9-294E-4C05-A660-B4F0D5C18AC5}
-	EndGlobalSection
-EndGlobal
