--- MongoDbAccess/CollectionImplementationAttribute.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace MongoDbAccess
-{
-	[AttributeUsage(AttributeTargets.Class)]
-	public class CollectionImplementationAttribute : Attribute
-	{
-	}
-}
