--- MongoDbAccess/DataAccess/Abstractions/IDbConnection.cs
@@ -1,12 +1,0 @@
-﻿using MongoDB.Driver;
-using MongoDbAccess.Models;
-
-namespace MongoDbAccess.DataAccess.Abstractions
-{
-	public interface IDbConnection
-	{
-		string GetCollectionName<TDocument>() where TDocument : IModel;
-
-		IMongoCollection<TDocument> GetCollection<TDocument>(string name) where TDocument : IModel;
-	}
-}
