--- MongoDbAccess/Models/IModel.cs
@@ -1,7 +1,0 @@
-﻿namespace MongoDbAccess.Models
-{
-	public interface IModel
-	{
-		public string Id { get; set; }
-	}
-}
