--- MongoDbAccess/MongoDbAccess.csproj
@@ -1,20 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <Version>2.0.0</Version>
-    <Authors>FK</Authors>
-    <Company>FK Family</Company>
-    <SignAssembly>true</SignAssembly>
-    <AssemblyOriginatorKeyFile>MongoDbAccess.snk</AssemblyOriginatorKeyFile>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageOutputPath>d:\code\dotnet\ref\</PackageOutputPath>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="7.0.0" />
-    <PackageReference Include="MongoDB.Driver" Version="2.18.0" />
-  </ItemGroup>
-
-</Project>
