+++ MongoExt.Generators/ModelSyntaxReceiver.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.CodeAnalysis;
+
+namespace MongoDbAccess.Generators
+{
+	internal class ModelSyntaxReceiver : ISyntaxReceiver
+	{
+        public ModelAttributeSyntaxReceiver ModelAttribute { get; } = new();
+        public CollectionDefinitionAttributeSyntaxReceiver CollectionDefinitionAttribute { get; } = new();
+
+        public void OnVisitSyntaxNode(SyntaxNode syntaxNode)
+        {
+            ModelAttribute.OnVisitSyntaxNode(syntaxNode);
+            CollectionDefinitionAttribute.OnVisitSyntaxNode(syntaxNode);
+        }
+    }
+}
