+++ MongoExt/CollectionDefinitionAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace MongoDbAccess
+{
+	[AttributeUsage(AttributeTargets.Interface)]
+	public class CollectionDefinitionAttribute : Attribute
+	{
+	}
+}
