+++ MongoExt/CollectionImplementationAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace MongoDbAccess
+{
+	[AttributeUsage(AttributeTargets.Class)]
+	public class CollectionImplementationAttribute : Attribute
+	{
+	}
+}
