+++ MongoExt/Models/IModel.cs
@@ -1,0 +1,7 @@
+﻿namespace MongoDbAccess.Models
+{
+	public interface IModel
+	{
+		public string Id { get; set; }
+	}
+}
